Stop excluding non-library when packaging
We currently exclude files that come from dependencies. This is great
for library files as this prevents dependencies from being packaged.

However, this doesn't work for cases where folks also add binaries
to non-lib locations, for example the tools directory. In those cases,
we don't want to exclude files that are coming from dependencies and
instead assume the developer did this intentional.

This fixes #14.