Ensure path to NuGet.exe is stable
We use NuGet.exe in two different contexts:

    (1) Restoring packages
    (2) In our implementation of NuProj

We currently use the same path to NuGet in both cases. This, however,
has the unfortunate side effect that opening the solution without
running a command line build first will not build.

The reason is that the NuGet.exe we use for package restore is only
retrieved when building from the command line -- the IDE will use a
different version of NuGet.exe.

In order to have a stable path we separate these two concerns. We'll
use a solution wide package reference to NuGet.CommandLine which is
the one we use for our implementation. This version of NuGet.exe is
guaranteed to be available -- regardless of whether a command line
build happened or not.

For package restore, we download the latest version of NuGet.exe from
http://nuget.org. However, to avoid confusion we don't place this one
anywhere in our source tree. Instead, we put it in the local app data
which also has the nice side effect that it will be shared between
different projects. The location is the same that ASP.NET vNext is
using.