Replace copying assets with path lookup
We're currently copying a whole bunch in order to allow our tests to
find assets. Not only is this not particularly efficient, it also has
the tendency to become fragile.

We've two types of assets:

1. Static. Those are part of the source tree and don't change, meaning
   they don't have to be produced before they can be used, for example,
   the scenario files.

2. Dynamic. Those are assets that we need to produce before we can use
   them, for example, the NuProj.Tasks.dll.

For static assets, we compute the project directory root and then do a
simple path lookup. For dynamic resources, we make sure that the test
project declares depending on those. Those will be then looked up from
the build output.