Add infrastructure to simplify testing
I think one very common class of tests will be to build a single
solution, retrieving the NuGet packages and asserting their structure.

Add a scenario class that, given a name, finds the solution in the
source tree, restores packages, builds it, asserts a successful build,
and returns the NuGet packages produced by it.