Don't build dependencies when building inside the IDE
We don't want to build in case we're performing a design time build
as we are expected to not produce any assets.

We also don't want to build the references in cases where we build
inside the IDE. The reason is that Visual Studio already built our
dependencies. Doing it again can regress performance. However, the
real issue is that it impacts correctness as this can result in
building the same project simultaneously from different projects.

Most particularly on the correctness side, this shows up when VS is
doing a "rebuild". NuProj will end up causing multiple build breaks
being reported because it will re-delete outputs that VS just produced
and that other project references that are building in parallel now
expect to be there.

This fixes issue #13.