Ensure that we don't package files coming from indirect dependencies
For example, consider this scenario:

  A.csproj => ANuGetPackage
  A.nuproj => A.csproj
  B.csproj => A.csproj
  B.nuproj => B.csproj, A.nuproj

B.nupkg should not contain assemblies from ANuGetPackage.

We're fixing this by injecting a target into Microsoft.Common.targets
that allows us to compute the closure of project references. We can
then use that to discover the closure of packages that come from NuGet
and are produced by NuProj projects.

This fixes issue #10.