Remove Input/Output specification from GenerateNuSpec
We always want to run this target; previously we made a fix to include
a fake input which results in always running it. We might as well just
remove the entire input/output specification and force the target to
always run.

In the end, it's the job of the GenerateNuSpec task to ensure it doesn't
change the timestamp if the file if the contents haven't been changed.

This also adds test for testing incremental build works as expected.

This should permanently fix #29.