Ensure the default item type is 'Content'
For a NuGet package project, the default item type should be 'Content',
rather than 'None'. In the end, folks use it to craft the contents of
their NuGet package

It's obviously still desirable to be able to exclude files from being
packaged. That's still possible as folks can simply change the content
type to 'None' if needed.