Update expected error message to match actual behavior
When we changed the NuSpec file generation we inadvertently changed the
values that are passed into NuGet.Core. Previously, the references and
and dependencies were passed in as an IEnumerable. Now they are passed
in as a List<T>.

Apparently, this causes NuGet.Core to go down different code paths when
checking whether the package is valid. The new path is actually better
because it results in a simpler, and more actionable error message:

     Cannot create package that has no dependencies nor content.

Thus, I've updated our test to assert the new error message.