Enable projects to determine the package version at build time.
When the version is determined based on the date, or the contents of a version.txt file, or otherwise simply that the version should be calculated at build time such as is specified in issue #20, we need to be able to set the $(Version) property after evaluation has completed.

This change ensures that an extensible set of Targets can be executed before $(Version) is ever used. So for example in my project I can now add the following build authoring to my nuproj file in order to calculate the package version:

    <VersionDependsOn>$(VersionDependsOn);GetNuPkgVersion</VersionDependsOn>

  <Target Name="GetNuPkgVersion" DependsOnTargets="GetBuildVersion">
    <PropertyGroup>
      <Version>$(NuGetPackageVersion)</Version>
    </PropertyGroup>
  </Target>