Xunit to display only test method names
Before this, Xunit2 would show fully-qualified test methods (with namespace and class) in its list, which is redundant with the test class name which has been embedded in each test method's own name.

Eventually the "right" fix is probably to simplify all our test method names and then revert this change. But that will be considerably more work, and in the meantime this commit will make the UI more accessible."