Update version from 0.9.3 to 0.10.0
The increment of the minor version rather than the patch version is based on the fact that significant and backward breaking changes have been made. Yes, the nuget package is "unstable" right now (which is why we're not obliged to increment the major component), but we should try to communicate to consumers that something more than a simple patch update has been applied. And the minor component seems like the best place to do that.