Add PrepareForBuild target
This is a popular target from Microsoft.Common.targets but NuProj didn't import or define it. This caused issues when a NuGet package with a .props or .targets import is installed into a NuProj itself, and the NuGet package manager adds a target to the NuProj that runs with PrepareForBuild, but since we didn't define the target, NuGet's own target never ran.