Stop inserting CustomAfterMicrosoftCommonTargets global property
This extra global property set on project references leads to building the referenced project multiple times. The solution builds it without that global property, so if NuProj builds it with the global property, MSBuild will build the project twice. When multi-proc build is enabled, this often leads to timing errors when the same project builds concurrently across processes and ends up with file access errors.